api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.1
python-multipart==0.0.9
pydantic-settings==2.4.0